requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
